# output/agent-interaction-contract.openapi.yaml
openapi: 3.1.0
info:
  title: Bittrees Agent Interaction Contract
  version: 1.0.1
  license:
    name: MIT
    identifier: MIT
  description: |
    OpenAPI projection of the shipped agent.bittrees.org discovery, identity,
    submission, and status routes. The contract is review-gated; public write
    routes may fail closed until the portal enables them explicitly. Canonical
    workflow routes use `/v1/workflow/*`; the server also rewrites the legacy
    `/v1/contributions/*` prefix to the same handlers.
...
servers:
  - url: https://agent.bittrees.org
security: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: >
        Pre-provisioned opaque bearer token resolved from the server's
        MCP_WRITE_TOKENS configuration. The implementation has no token issue,
        exchange, or refresh endpoint. Configured status, not-before, expiry,
        and revocation metadata is checked when present; each write also checks
        its operation scope and applicable token-subject binding.
  parameters:
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: false
      description: >
        Optional retry key. The portal also accepts a request-body fallback
        (`intentId` for contribution-intents or `idempotencyKey` for workflow
        mutations).
      schema:
        type: string
        minLength: 1
        maxLength: 180
        pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,179}$'
...
  requestBodies:
    ContributionIntentSubmission:
      required: true
      description: >
        Contribution-intent payload accepted as JSON or URL-encoded form data.
        The gateway form path uses the same normalized request object after
        parsing dotted field names and repeated array fields.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ContributionIntentRequest'
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/ContributionIntentRequest'
...
  schemas:
...
    ContributionIntentContractResponse:
      $ref: './schemas/contribution-intent-contract-response.v1.json'
    ContributionIntentRequest:
      $ref: './schemas/contribution-intent-request.v1.json'
    ContributionIntentResponse:
      $ref: './schemas/contribution-intent-response.v1.json'
...
    ErrorEnvelope:
      $ref: './schemas/error-envelope.v1.json'
...
  responses:
...
    ContributionIntentContract:
      description: Contribution-intent contract response for discovery and form guidance.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ContributionIntentContractResponse'
    ContributionIntentIntake:
      description: >
        Contribution-intent intake response envelope. The same schema is used
        for accepted, rejected, rate-limited, and write-disabled outcomes.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ContributionIntentResponse'
    ErrorEnvelope:
      description: Structured error envelope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
paths:
...
  /contribution-intents:
    get:
      tags:
        - submission
      operationId: getContributionIntentContract
      summary: Get the contribution-intent contract
      description: >
        Machine-readable contribution-intent contract and launch posture for
        the intake route. The response includes the public privacy notice and
        form submission contract.
      responses:
        '200':
          $ref: '#/components/responses/ContributionIntentContract'
        '400':
          $ref: '#/components/responses/ErrorEnvelope'
    post:
      tags:
        - submission
      operationId: submitContributionIntent
      summary: Submit a contribution-intent packet
      description: >
        Contribution-intent intake is review-gated. Production currently fails
        closed with 501 unless a non-production write flag is enabled. When
        live writes are enabled, the route authenticates with a bearer token
        carrying the `contributor:submit` scope.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        $ref: '#/components/requestBodies/ContributionIntentSubmission'
      responses:
        '202':
          $ref: '#/components/responses/ContributionIntentIntake'
        '400':
          $ref: '#/components/responses/ContributionIntentIntake'
        '401':
          $ref: '#/components/responses/ContributionIntentIntake'
        '403':
          $ref: '#/components/responses/ContributionIntentIntake'
        '409':
          $ref: '#/components/responses/ContributionIntentIntake'
        '415':
          $ref: '#/components/responses/ContributionIntentIntake'
        '422':
          $ref: '#/components/responses/ContributionIntentIntake'
        '429':
          $ref: '#/components/responses/ContributionIntentIntake'
        '501':
          $ref: '#/components/responses/ContributionIntentIntake'
  /gateway/contribution-intents:
    get:
      tags:
        - submission
      operationId: getGatewayContributionIntentContract
      summary: Get the gateway contribution-intent contract
      description: >
        Browser-oriented view of the same contribution-intent contract and form
        submission rules. The response mirrors the JSON contract route.
      responses:
        '200':
          $ref: '#/components/responses/ContributionIntentContract'
        '400':
          $ref: '#/components/responses/ErrorEnvelope'
    post:
      tags:
        - submission
      operationId: submitGatewayContributionIntent
      summary: Submit a contribution-intent packet through the gateway form
      description: >
        Browser form submissions hit the same review-gated intake pipeline as
        the JSON route. Form posts may be encoded as
        application/x-www-form-urlencoded with the same normalized fields.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        $ref: '#/components/requestBodies/ContributionIntentSubmission'
      responses:
        '202':
          $ref: '#/components/responses/ContributionIntentIntake'
        '400':
          $ref: '#/components/responses/ContributionIntentIntake'
        '401':
          $ref: '#/components/responses/ContributionIntentIntake'
        '403':
          $ref: '#/components/responses/ContributionIntentIntake'
        '409':
          $ref: '#/components/responses/ContributionIntentIntake'
        '415':
          $ref: '#/components/responses/ContributionIntentIntake'
        '422':
          $ref: '#/components/responses/ContributionIntentIntake'
        '429':
          $ref: '#/components/responses/ContributionIntentIntake'
        '501':
          $ref: '#/components/responses/ContributionIntentIntake'
...