# scripts/serve-dist.mjs
import { createServer } from 'node:http';
import { readFile } from 'node:fs/promises';
import { extname, join, normalize, relative, sep } from 'node:path';
import { fileURLToPath } from 'node:url';

import { MCP_GATEWAY, PORTAL_SECURITY_HEADERS, ROBOTS_TXT_PATH, ROUTE_DEFINITIONS, createRequestHandler, handleMcpRequest, normalizeCanonicalPath } from '../src/portal.mjs';

const rootDir = fileURLToPath(new URL('..', import.meta.url));
const distDir = join(rootDir, 'dist');
const port = Number(process.env.PORT ?? '3000');
const host = process.env.HOST ?? '0.0.0.0';
const canonicalRoutePaths = new Set([ROBOTS_TXT_PATH, ...ROUTE_DEFINITIONS.map((definition) => definition.path), '/portal-manifest.json']);
const routeDefinitionsByPath = new Map(ROUTE_DEFINITIONS.map((definition) => [definition.path, definition]));
const dynamicPortalHandler = createRequestHandler();
// These routes must execute the server handler so query parameters reach the
// contribution service projection loader. Serving dist/submission-status as a
// static snapshot would make every lookup appear identical.
const dynamicPortalRoutePaths = new Set([
  '/contribution-intents',
  '/gateway/contribution-intents',
  '/submission-status',
  '/submission-status/index.html',
]);
// API namespaces are implemented only by createRequestHandler. Delegate their
// full namespaces rather than enumerating today's endpoints, so dist-mode
// stays aligned with source-mode as the workflow and registry contracts grow.
const dynamicPortalRoutePrefixes = ['/v1/workflow/', '/v1/registry/'];

function isDynamicPortalPath(pathname) {
  return dynamicPortalRoutePaths.has(pathname)
    || dynamicPortalRoutePrefixes.some((prefix) => pathname.startsWith(prefix));
}
const extensionlessStaticRoutePaths = new Set(
  ROUTE_DEFINITIONS
    .filter((definition) => definition.kind !== 'html' && definition.staticAsset !== false && !extname(definition.path))
    .map((definition) => definition.path),
);

// Every extension the dist server can emit needs an explicit type here. The
// static and error responses set `X-Content-Type-Options: nosniff`, so the
// browser will not infer a type: an SVG icon or an XML sitemap served as the
// `application/octet-stream` fallback is downloaded or refused instead of
// rendered. Keep these aligned with the asset kinds scripts/build.mjs can emit.
const contentTypes = new Map([
  ['.html', 'text/html; charset=utf-8'],
  ['.json', 'application/json; charset=utf-8'],
  ['.txt', 'text/plain; charset=utf-8'],
  ['.svg', 'image/svg+xml; charset=utf-8'],
  ['.xml', 'application/xml; charset=utf-8'],
  ['.png', 'image/png'],
]);

function resolveAssetPath(requestUrl) {
  const url = new URL(requestUrl ?? '/', 'http://localhost');
  const normalizedRoutePath = normalizeCanonicalPath(url.pathname);
  const pathname = url.pathname === '/'
    ? '/index.html'
    : extensionlessStaticRoutePaths.has(normalizedRoutePath)
      ? normalizedRoutePath
      : extname(url.pathname)
        ? url.pathname
        : `${url.pathname.replace(/\/$/, '')}/index.html`;
  let decodedPathname;

  try {
    decodedPathname = decodeURIComponent(pathname);
  } catch {
    return null;
  }

  const normalizedPath = normalize(decodedPathname).replace(/^(\.\.[/\\])+/, '');
  const assetPath = join(distDir, normalizedPath);
  const relativePath = relative(distDir, assetPath);

  if (relativePath.startsWith('..') || relativePath.includes(`..${sep}`)) {
    return null;
  }

  return assetPath;
}

function getContentType(assetPath, requestPathname) {
  const routeDefinition = routeDefinitionsByPath.get(normalizeCanonicalPath(requestPathname));

  if (routeDefinition?.kind === 'json') {
    return 'application/json; charset=utf-8';
  }

  return contentTypes.get(extname(assetPath)) ?? 'application/octet-stream';
}

function sendText(res, statusCode, body) {
  res.writeHead(statusCode, {
    'Content-Type': 'text/plain; charset=utf-8',
    'Content-Length': Buffer.byteLength(body),
    'Cache-Control': 'no-store',
    'X-Content-Type-Options': 'nosniff',
    'X-Robots-Tag': 'noindex, nofollow',
    ...PORTAL_SECURITY_HEADERS,
  });
  res.end(body);
}

function logTelemetry({ timestamp = new Date().toISOString(), method, path, status }) {
  console.log(JSON.stringify({ timestamp, method, path, status }));
}

function sendRedirect(res, statusCode, location, telemetry) {
  res.writeHead(statusCode, {
    Location: location,
    'Content-Type': 'text/plain; charset=utf-8',
    'Content-Length': '0',
    'Cache-Control': 'no-store',
    'X-Content-Type-Options': 'nosniff',
    'X-Robots-Tag': 'noindex, nofollow',
    ...PORTAL_SECURITY_HEADERS,
  });
  res.end();
  logTelemetry(telemetry);
}

const server = createServer(async (req, res) => {
  const url = new URL(req.url ?? '/', 'http://localhost');
  const pathname = url.pathname;
  const telemetry = { method: req.method ?? 'GET', path: pathname };

  const normalizedPath = normalizeCanonicalPath(pathname);
  if (pathname !== normalizedPath && canonicalRoutePaths.has(normalizedPath)) {
    return sendRedirect(res, 301, `${normalizedPath}${url.search}`, {
      ...telemetry,
      status: 301,
    });
  }

  if (normalizedPath === MCP_GATEWAY.path) {
    return handleMcpRequest(req, res, telemetry).catch((error) => {
      console.error(error);
      if (!res.headersSent) {
        res.once('finish', () => logTelemetry({ ...telemetry, status: 500 }));
        sendText(res, 500, 'Internal MCP gateway error.\n');
      }
    });
  }

  if (isDynamicPortalPath(normalizedPath)) {
    return dynamicPortalHandler(req, res);
  }

  if (req.method !== 'GET' && req.method !== 'HEAD') {
    res.once('finish', () => logTelemetry({ ...telemetry, status: 405 }));
    return sendText(res, 405, 'Only GET and HEAD are supported.\n');
  }

  const assetPath = resolveAssetPath(req.url);

  if (!assetPath) {
    res.once('finish', () => logTelemetry({ ...telemetry, status: 404 }));
    return sendText(res, 404, 'Not found.\n');
  }

  try {
    const body = await readFile(assetPath);
    res.once('finish', () => logTelemetry({ ...telemetry, status: 200 }));
    res.writeHead(200, {
      'Content-Type': getContentType(assetPath, pathname),
      'Content-Length': body.byteLength,
      'Cache-Control': 'no-store',
      'X-Content-Type-Options': 'nosniff',
      'X-Robots-Tag': 'noindex, nofollow',
      ...PORTAL_SECURITY_HEADERS,
    });
    res.end(req.method === 'HEAD' ? undefined : body);
  } catch (error) {
    if (error.code === 'ENOENT' || error.code === 'EISDIR') {
      // Let the portal handler content-negotiate unknown routes: browser GETs
      // receive the branded HTML 404 while API and agent clients retain the
      // structured JSON error contract. Static assets remain directly readable.
      return dynamicPortalHandler(req, res);
    }

    console.error(error);
    res.once('finish', () => logTelemetry({ ...telemetry, status: 500 }));
    return sendText(res, 500, 'Internal server error.\n');
  }
});

server.listen(port, host, () => {
  console.log(`serving ${relative(rootDir, distDir)}/ on http://${host}:${port}`);
});

function shutdown(signal) {
  server.close(() => {
    console.log(`received ${signal}, stopped cleanly`);
    process.exit(0);
  });
}

process.on('SIGINT', () => shutdown('SIGINT'));
process.on('SIGTERM', () => shutdown('SIGTERM'));